VOIP-8029: staging: brcmesw: Use fixed size types for esw_reg_data
Change the 'len' field from size_t to a fixed size uint8_t.
size_t will vary in size depending on the compiler.
Since this file is shared with user space if we build user
space with 32 bit compiler and the kernel is 64 bit then
we get mismatched structures. This causes problems when
communicating the data to the kernel.

Reviewed-on: http://gerrit-ccxsw.broadcom.net/79537
Tested-by: Scott Branden <scott.branden@broadcom.com>

+++ include/uapi/linux/bcm_esw_ioctl.h
@@ -14,7 +14,7 @@ struct esw_reg_data {
 	/* Switch access using 64-bit only */
 	uint64_t data;
 	/* Actually register length, in bits */
-	size_t len;
+	uint8_t len;
 	uint8_t page;
 	uint8_t offset;
 };
